requirements.txt
flask
pandas
numpy
sentence-transformers
faiss-cpu
plotly
python-dotenv
requests
pypdf
google-genai
pymongo